cloud_io: adopt cloud_storage::materialized_resources
Moves the throttling code of materialized_resources into a new
cloud_io::io_resources. This will be useful for other projects that need
throttling that don't care about tracking segments (which is also done
by materialized_resources).
