+++ src/v/CMakeLists.txt
@@ -129,6 +129,7 @@ add_subdirectory(rp_util)
 add_subdirectory(resource_mgmt)
 add_subdirectory(migrations)
 add_subdirectory(datalake)
+add_subdirectory(cloud_io)
 
 option(ENABLE_GIT_VERSION "Build with Git metadata" OFF)
