+++ src/v/cloud_io/CMakeLists.txt
@@ -1,0 +1,8 @@
+v_cc_library(
+  NAME cloud_io
+  SRCS
+    io_resources.cc
+  DEPS
+    Seastar::seastar
+    v::config)
+
