+++ src/v/cloud_storage/materialized_resources.cc
@@ -41,123 +41,11 @@ using namespace std::chrono_literals;
 static constexpr ss::lowres_clock::duration stm_jitter_duration = 10s;
 static constexpr ss::lowres_clock::duration stm_max_idle_time = 60s;
 
-struct device_throughput {
-    size_t read = std::numeric_limits<size_t>::max();
-    size_t write = std::numeric_limits<size_t>::max();
-};
-
-struct throughput_limit {
-    size_t disk_node_throughput_limit = std::numeric_limits<size_t>::max();
-    size_t download_shard_throughput_limit = std::numeric_limits<size_t>::max();
-};
-
-/// Compute integer multiplication and division without the integer overflow
-/// \returns val * mul / div
-inline size_t muldiv(size_t val, size_t mul, size_t div) {
-    const auto dv = std::lldiv(
-      static_cast<long long>(val), static_cast<long long>(div));
-    return dv.quot * mul + dv.rem * mul / div;
-}
-
-// Get device throughput for the mountpoint
-inline ss::future<std::optional<device_throughput>>
-get_storage_device_throughput() {
-    if (config::shard_local_cfg().cloud_storage_enabled() == false) {
-        co_return std::nullopt;
-    }
-    try {
-        auto cache_path = config::node().cloud_storage_cache_path().native();
-        auto fs = co_await ss::file_stat(cache_path);
-        auto& queue = ss::engine().get_io_queue(fs.device_id);
-        auto cfg = queue.get_config();
-        auto percent = config::shard_local_cfg()
-                         .cloud_storage_throughput_limit_percent()
-                         .value_or(0);
-        if (percent > 0) {
-            // percent == nullopt indicates that the throttling is disabled
-            // intentionally
-            co_return device_throughput{
-              .read = muldiv(cfg.read_bytes_rate, percent, 100),
-              .write = muldiv(cfg.write_bytes_rate, percent, 100),
-            };
-        }
-    } catch (...) {
-        vlog(
-          cst_log.info,
-          "Can't get device throughput: {} for {} mountpoint",
-          std::current_exception(),
-          config::node().cloud_storage_cache_path().native());
-    }
-    co_return std::nullopt;
-}
-
-// Compute device limit based on configuration. The value is a total throughput
-// for the node. Doesn't take into account hardware configuration.
-inline throughput_limit get_hard_throughput_limit() {
-    auto hard_limit = config::shard_local_cfg()
-                        .cloud_storage_max_throughput_per_shard()
-                        .value_or(0)
-                      * ss::smp::count;
-
-    if (hard_limit == 0) {
-        // Run tiered-storage without throttling by setting
-        // 'cloud_storage_max_throughput_per_shard' to nullopt
-        return {};
-    }
-
-    return {
-      .disk_node_throughput_limit = hard_limit,
-      .download_shard_throughput_limit = hard_limit / ss::smp::count,
-    };
-}
-
-// Compute device limit based on configuration. The value is a total throughput
-// for the node. This function takes hardware into account.
-inline throughput_limit
-get_throughput_limit(std::optional<size_t> device_throughput) {
-    auto hard_limit = config::shard_local_cfg()
-                        .cloud_storage_max_throughput_per_shard()
-                        .value_or(0)
-                      * ss::smp::count;
-
-    if (
-      config::shard_local_cfg()
-          .cloud_storage_throughput_limit_percent()
-          .value_or(0)
-        == 0
-      || hard_limit == 0) {
-        // Run tiered-storage without throttling by setting
-        // 'cloud_storage_throughput_limit_percent' to nullopt or
-        // 'cloud_storage_max_throughput_per_shard' to nullopt
-        return throughput_limit{};
-    }
-
-    if (!device_throughput.has_value()) {
-        // The 'cloud_storage_throughput_limit_percent' is set
-        // but but we couldn't read the actual device throughput. No need
-        // to limit the disk bandwidth in this case. But since hard limit
-        // is set we still need to limit network bandwidth even though
-        // the limit is overly high.
-        return throughput_limit{
-          .download_shard_throughput_limit = hard_limit / ss::smp::count,
-        };
-    }
-
-    auto tp = std::min(hard_limit, device_throughput.value());
-    return {
-      .disk_node_throughput_limit = tp,
-      .download_shard_throughput_limit = tp / ss::smp::count,
-    };
-}
-
 materialized_resources::materialized_resources()
   : _stm_jitter(stm_jitter_duration)
   , _max_segment_readers_per_shard(
       config::shard_local_cfg()
         .cloud_storage_max_segment_readers_per_shard.bind())
-  , _max_concurrent_hydrations_per_shard(
-      config::shard_local_cfg()
-        .cloud_storage_max_concurrent_hydrations_per_shard.bind())
   , _storage_read_buffer_size(
       config::shard_local_cfg().storage_read_buffer_size.bind())
   , _storage_read_readahead_count(
@@ -165,19 +53,10 @@ materialized_resources::materialized_resources()
   , _mem_units(
       memory_groups().tiered_storage_max_memory(),
       "cst_materialized_resources_memory")
-  , _hydration_units(max_parallel_hydrations(), "cst_hydrations")
   , _manifest_meta_size(
       config::shard_local_cfg().cloud_storage_manifest_cache_size.bind())
   , _manifest_cache(ss::make_shared<materialized_manifest_cache>(
       config::shard_local_cfg().cloud_storage_manifest_cache_size()))
-  , _throughput_limit(
-      // apply shard limit to downloads
-      get_hard_throughput_limit().download_shard_throughput_limit,
-      "ts-segment-downloads")
-  , _throughput_shard_limit_config(
-      config::shard_local_cfg().cloud_storage_max_throughput_per_shard.bind())
-  , _relative_throughput(
-      config::shard_local_cfg().cloud_storage_throughput_limit_percent.bind())
   , _cache_carryover_bytes(config::shard_local_cfg()
                              .cloud_storage_cache_trim_carryover_bytes.bind()) {
     auto update_max_mem = [this]() {
@@ -188,11 +67,6 @@ materialized_resources::materialized_resources()
     _max_segment_readers_per_shard.watch(update_max_mem);
     _storage_read_buffer_size.watch(update_max_mem);
     _storage_read_readahead_count.watch(update_max_mem);
-    _max_concurrent_hydrations_per_shard.watch([this]() {
-        // The 'max_connections' parameter can't be changed without restarting
-        // redpanda.
-        _hydration_units.set_capacity(max_parallel_hydrations());
-    });
 
     _manifest_meta_size.watch([this] {
         ssx::background = ss::with_gate(_gate, [this] {
@@ -250,68 +124,6 @@ materialized_resources::materialized_resources()
               _carryover_units.has_value() ? _carryover_units->count() : 0);
         });
     }
-
-    auto reset_tp = [this] {
-        ssx::spawn_with_gate(_gate, [this] { return update_throughput(); });
-    };
-
-    _throughput_shard_limit_config.watch(reset_tp);
-    _relative_throughput.watch(reset_tp);
-
-    reset_tp();
-}
-
-ss::future<> materialized_resources::update_throughput() {
-    auto tp = get_throughput_limit(_device_throughput);
-    if (ss::this_shard_id() == 0) {
-        co_await set_disk_max_bandwidth(tp.disk_node_throughput_limit);
-    }
-    set_net_max_bandwidth(tp.download_shard_throughput_limit);
-}
-
-void materialized_resources::set_net_max_bandwidth(size_t tput) {
-    if (tput != 0 && tput != std::numeric_limits<size_t>::max()) {
-        vlog(
-          cst_log.info,
-          "Setting cloud storage download bandwidth to {} on this shard",
-          tput);
-        _throughput_limit.update_capacity(tput);
-        _throughput_limit.update_rate(tput);
-        _throttling_disabled = false;
-    } else {
-        vlog(
-          cst_log.info,
-          "Disabling cloud storage download throttling on this shard");
-        _throttling_disabled = true;
-    }
-}
-
-ss::future<> materialized_resources::set_disk_max_bandwidth(size_t tput) {
-    try {
-        if (tput == 0 || tput == std::numeric_limits<size_t>::max()) {
-            _throttling_disabled = true;
-            vlog(
-              cst_log.info,
-              "Scheduling group's {} bandwidth is not limited",
-              priority_manager::local().shadow_indexing_priority().get_name());
-            co_return;
-        }
-        _throttling_disabled = false;
-        vlog(
-          cst_log.info,
-          "Setting scheduling group {} bandwidth to {}",
-          priority_manager::local().shadow_indexing_priority().get_name(),
-          tput);
-
-        co_await priority_manager::local()
-          .shadow_indexing_priority()
-          .update_bandwidth(tput);
-    } catch (...) {
-        vlog(
-          cst_log.error,
-          "Failed to set tiered-storage disk throughput: {}",
-          std::current_exception());
-    }
 }
 
 ts_read_path_probe& materialized_resources::get_read_path_probe() {
@@ -321,15 +133,11 @@ ts_read_path_probe& materialized_resources::get_read_path_probe() {
 ss::future<> materialized_resources::stop() {
     cst_log.debug("Stopping materialized_segments...");
 
-    _throughput_limit.shutdown();
-
     co_await _manifest_cache->stop();
 
     _stm_timer.cancel();
 
     _mem_units.broken();
-
-    _hydration_units.broken();
 
     co_await _gate.close();
     cst_log.debug("Stopped materialized_segments...");
@@ -344,12 +152,6 @@ ss::future<> materialized_resources::start() {
     _stm_timer.rearm(_stm_jitter());
 
     co_await _manifest_cache->start();
-
-    auto tput = co_await get_storage_device_throughput();
-    if (tput.has_value()) {
-        _device_throughput = tput->write;
-    }
-    co_await update_throughput();
 
     co_return;
 }
@@ -398,28 +200,12 @@ size_t materialized_resources::max_memory_utilization() const {
     return memory_groups().tiered_storage_max_memory();
 }
 
-size_t materialized_resources::max_parallel_hydrations() const {
-    auto max_connections
-      = config::shard_local_cfg().cloud_storage_max_connections();
-    auto n_readers = config::shard_local_cfg()
-                       .cloud_storage_max_partition_readers_per_shard();
-    if (n_readers) {
-        return std::min(
-          static_cast<unsigned>(max_connections), n_readers.value());
-    }
-    return max_connections / 2;
-}
-
 size_t materialized_resources::current_segment_readers() const {
     size_t n = 0;
     for (const auto& s : _materialized) {
         n += s.readers.size();
     }
     return n;
-}
-
-size_t materialized_resources::current_ongoing_hydrations() const {
-    return _hydration_units.outstanding();
 }
 
 size_t materialized_resources::current_segments() const {
@@ -506,12 +292,6 @@ materialized_resources::get_partition_reader_units(
     }
 
     return get_units_abortable(_mem_units, sz, as);
-}
-
-ss::future<ssx::semaphore_units>
-materialized_resources::get_hydration_units(size_t n) {
-    auto u = co_await _hydration_units.get_units(n);
-    co_return std::move(u);
 }
 
 ss::future<segment_units>
@@ -715,78 +495,6 @@ void materialized_resources::maybe_trim_segment(
             st.readers.pop_front();
         }
     }
-}
-
-class throttled_dl_source : public ss::data_source_impl {
-public:
-    explicit throttled_dl_source(
-      ss::input_stream<char> st,
-      materialized_resources& parent,
-      ss::abort_source& as,
-      ss::gate::holder h)
-      : _src(std::move(st).detach())
-      , _parent(parent)
-      , _as(as)
-      , _holder(std::move(h)) {}
-
-    ss::future<ss::temporary_buffer<char>> get() override {
-        auto h = _gate.hold();
-        return _src.get().then(
-          [this, h = std::move(h)](ss::temporary_buffer<char> buf) mutable {
-              size_t buffer_size = buf.size();
-              return maybe_throttle(buffer_size)
-                .then([b = std::move(buf), h = std::move(h)]() mutable {
-                    return std::move(b);
-                });
-          });
-    }
-
-    ss::future<ss::temporary_buffer<char>> skip(uint64_t n) override {
-        auto holder = _gate.hold();
-        return _src.skip(n).finally(
-          [this, n, h = std::move(holder)] { return maybe_throttle(n); });
-    }
-
-    ss::future<> close() override {
-        _holder.release();
-        return _gate.close().then([this] { return _src.close(); });
-    }
-
-private:
-    ss::future<> maybe_throttle(size_t buffer_size) {
-        auto throttled = co_await _parent._throughput_limit.maybe_throttle(
-          buffer_size, _as);
-        if (throttled) {
-            // This path is taken only when the download is throttled. In case
-            // of concurrency every `maybe_throttle` call will return the
-            // precise value for deficiency and sleep time.
-            auto duration
-              = std::chrono::duration_cast<std::chrono::milliseconds>(
-                throttled.value());
-            _parent._read_path_probe.download_throttled(duration.count());
-            vlog(
-              cst_log.trace,
-              "Download throttled: sleep time is {} ms",
-              duration.count());
-        }
-    }
-    ss::data_source _src;
-    materialized_resources& _parent;
-    ss::gate _gate;
-    ss::abort_source& _as;
-    // Gate holder of the `_parent`
-    ss::gate::holder _holder;
-};
-
-ss::input_stream<char> materialized_resources::throttle_download(
-  ss::input_stream<char> underlying, ss::abort_source& as) {
-    if (_throttling_disabled) {
-        return underlying;
-    }
-    auto src = std::make_unique<throttled_dl_source>(
-      std::move(underlying), *this, as, _gate.hold());
-    ss::data_source ds(std::move(src));
-    return ss::input_stream<char>(std::move(ds));
 }
 
 } // namespace cloud_storage
