+++ src/v/cloud_storage/remote_probe.h
@@ -11,6 +11,7 @@
 #pragma once
 
 #include "base/seastarx.h"
+#include "cloud_io/io_resources.h"
 #include "cloud_storage/types.h"
 #include "metrics/metrics.h"
 #include "model/fundamental.h"
@@ -32,7 +33,8 @@ public:
     explicit remote_probe(
       remote_metrics_disabled disabled,
       remote_metrics_disabled public_disabled,
-      materialized_resources&);
+      materialized_resources&,
+      const cloud_io::io_resources&);
     remote_probe(const remote_probe&) = delete;
     remote_probe& operator=(const remote_probe&) = delete;
     remote_probe(remote_probe&&) = delete;
